Precision should be float not int and remove default on abstract

+++ src/Troward/Converter/Measurement/Area.php
@@ -15,9 +15,9 @@ class Area extends Measurement
      *
      * @param float $input
      * @param string $unit
-     * @param int $precision
+     * @param float $precision
      */
-    public function __construct(float $input, string $unit, int $precision = 2)
+    public function __construct(float $input, string $unit, float $precision = 2)
     {
         $this->validateUnit(AreaUnit::class, $unit);
         parent::__construct($input, $unit, $precision);
